requirements.txt
nose
pygame
deprecated
pillow